+++ extras/blur_prepare.py
@@ -5,7 +5,7 @@ import time
 from PIL import Image as Im
 from PIL import ImageFilter
 
-from extras import extrasUtil
+from utils import util
 
 # Helper Variables
 slash = "\\" if os.name == 'nt' else "/"
@@ -43,7 +43,7 @@ def get_random_blur_type():
 
 
 def process(input_folder):
-    file_count = extrasUtil.check_file_count(input_folder)
+    file_count = util.check_file_count(input_folder)
     index = 1
     rgb_index = 0
     failed_files = 0
