+++ extras/luma_prepare.py
@@ -6,7 +6,7 @@ from PIL import Image as Im
 from PIL import ImageCms as ImCms
 from PIL import ImageFilter
 
-from extras import extrasUtil
+from utils import util
 
 # Helper Variables
 
@@ -39,7 +39,7 @@ def get_random_radius(a, b):
 
 
 def process(input_folder):
-    file_count = extrasUtil.check_file_count(input_folder)
+    file_count = util.check_file_count(input_folder)
     index = 1
     failed_files = 0
     skipped_files = 0
